refactor(oauth): extract RedirectPkceOauthBaseService
Pulls the loopback callback server, PKCE state, token exchange/refresh,
and revoke logic out of OpenaiOauthService into a reusable abstract base
keyed by a config object (provider id, URLs, scope, callback port,
optional client secret, extra authorize params).

OpenaiOauthService becomes a 16-line subclass that supplies the OpenAI
config and inherits all behavior. Public API and existing test contract
preserved verbatim — all 55 OpenAI OAuth specs pass without changes.

Motivated by adding Gemini OAuth (gemini-cli flow): the second
redirect-PKCE provider would otherwise duplicate ~290 lines of identical
plumbing.
